requirements.txt
streamlit>=1.39
python-dotenv>=1.0
google-cloud-firestore>=2.16
firebase-admin>=6.5
pydantic>=2.7
requests>=2.31